Set up the source tree for CodeMirror Speak-and-Spell.
This sets up the build environment, package information, and public
information for CodeMirror Speak-and-Spell, a plugin for CodeMirror v5
that allows reliable usage of browser-provided spell checking,
speech-to-text, AI assistated writing, and more.

There's no implementation of the plugin in this change, but that will
come next.

Testing Done:
Built this along with the plugin code and tested it in Firefox, Chrome,
and Safari.

Reviewed at https://reviews.reviewboard.org/r/14474/